Safari-611.1.10/LayoutTests/fast/events/scale-and-scroll-iframe-window-expected.txt

PASS frame.contentDocument.body.scrollTop is 100
PASS frame.contentDocument.body.scrollLeft is 100
PASS frame.contentWindow.scrollX is 100
PASS frame.contentWindow.scrollY is 100
